update Player variable type for validating game play

+++ ArrowStorm/core/src/com/ayp/arrowstormgame/model/Enemy.java
@@ -70,7 +70,7 @@ public abstract class Enemy {
         return healthPoint <= 0;
     }
 
-    public void takeDamage(int damage) {
+    public void takeDamage(float damage) {
         healthPoint -= damage;
     }
